requirements.txt
asgiref==3.8.1
bidict==0.23.1
blinker==1.9.0
channels==4.2.2
click==8.1.8
colorama==0.4.6
Django==5.2
dnspython==2.7.0
eventlet==0.40.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
pyserial==3.5
python-decouple==3.8
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
sqlparse==0.5.3
tzdata==2025.2
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
